# assets/generated/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cluster Manager AKS Helper API",
    "version": "v0.1.0",
    "contact": {
      "name": "Michael Venezia",
      "url": "https://github.com/samsung-cnct/cma-aks",
      "email": "mvenezia@gmail.com"
    }
  },
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/cluster": {
      "get": {
        "summary": "Will retrieve the status of a cluster and its kubeconfig for connectivity",
        "operationId": "GetCluster",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksGetClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Name of the cluster to be looked up.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.app_id",
            "description": "The AppId for API Access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.tenant",
            "description": "The Tenant for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.password",
            "description": "The Password for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.subscription_id",
            "description": "The Subscription for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "delete": {
        "summary": "Will delete a cluster",
        "operationId": "DeleteCluster",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksDeleteClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "What is the cluster's name to destroy.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.app_id",
            "description": "The AppId for API Access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.tenant",
            "description": "The Tenant for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.password",
            "description": "The Password for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.subscription_id",
            "description": "The Subscription for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "put": {
        "summary": "Will provision a cluster",
        "operationId": "CreateCluster",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksCreateClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmaaksCreateClusterMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
    "/api/v1/cluster/autoscaling": {
      "put": {
        "summary": "Will enable cluster autoscaling",
        "operationId": "EnableClusterAutoscaling",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksEnableClusterAutoscalingReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmaaksEnableClusterAutoscalingMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
    "/api/v1/cluster/list": {
      "get": {
        "summary": "Will retrieve a list of clusters",
        "operationId": "GetClusterList",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksGetClusterListReply"
            }
          }
        },
        "parameters": [
          {
            "name": "credentials.app_id",
            "description": "The AppId for API Access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.tenant",
            "description": "The Tenant for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.password",
            "description": "The Password for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.subscription_id",
            "description": "The Subscription for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
    "/api/v1/cluster/scale": {
      "get": {
        "summary": "Will retrieve node count",
        "operationId": "GetClusterNodeCount",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksGetClusterNodeCountReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Name of the cluster to be looked up.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.app_id",
            "description": "The AppId for API Access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.tenant",
            "description": "The Tenant for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.password",
            "description": "The Password for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.subscription_id",
            "description": "The Subscription for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "put": {
        "summary": "Will scale a clusters node count",
        "operationId": "ScaleCluster",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksScaleClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmaaksScaleClusterMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
    "/api/v1/cluster/upgrade": {
      "get": {
        "summary": "Will retrieve available upgrades of a cluster",
        "operationId": "GetClusterUpgrades",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksGetClusterUpgradesReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Name of the cluster to be looked up.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.app_id",
            "description": "The AppId for API Access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.tenant",
            "description": "The Tenant for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.password",
            "description": "The Password for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.subscription_id",
            "description": "The Subscription for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "put": {
        "summary": "Will upgrade a cluster",
        "operationId": "UpgradeCluster",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmaaksUpgradeClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmaaksUpgradeClusterMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
...
  },
  "definitions": {
...
    "cmaaksCreateClusterMsg": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Name of the cluster to be provisioned"
        },
        "provider": {
          "$ref": "#/definitions/cmaaksCreateClusterProviderSpec",
          "title": "The provider specification"
        }
      }
    },
...
    "cmaaksCreateClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Whether or not the cluster was provisioned by this request"
        },
        "cluster": {
          "$ref": "#/definitions/cmaaksClusterItem",
          "title": "The details of the cluster request response"
        }
      }
    },
    "cmaaksDeleteClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Could the cluster be destroyed"
        },
        "status": {
          "type": "string",
          "title": "Status of the request"
        }
      }
    },
    "cmaaksEnableClusterAutoscalingMsg": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Name of the cluster"
        },
        "min_quantity": {
          "type": "integer",
          "format": "int32",
          "title": "Minimum number of nodes for autoscaling"
        },
        "max_quantity": {
          "type": "integer",
          "format": "int32",
          "title": "Maximum number of nodes for autoscaling"
        },
        "credentials": {
          "$ref": "#/definitions/cmaaksAzureCredentials",
          "title": "Credentials to access the cluster"
        }
      }
    },
    "cmaaksEnableClusterAutoscalingReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Whether or not cluster autoscaling was enabled by this request"
        }
      }
    },
    "cmaaksGetClusterListReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Is the cluster in the system"
        },
        "clusters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/cmaaksClusterItem"
          },
          "title": "List of clusters"
        }
      }
    },
    "cmaaksGetClusterNodeCountReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Is the cluster in the system"
        },
        "name": {
          "type": "string",
          "title": "the available node pool name"
        },
        "count": {
          "type": "integer",
          "format": "int32",
          "title": "The available node count"
        }
      },
      "title": "Reply with available node count"
    },
    "cmaaksGetClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Is the cluster in the system"
        },
        "cluster": {
          "$ref": "#/definitions/cmaaksClusterDetailItem"
        }
      }
    },
...
    "cmaaksScaleClusterMsg": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Name of cluster"
        },
        "credentials": {
          "$ref": "#/definitions/cmaaksAzureCredentials",
          "title": "Credentials to search subscription for clusters"
        },
        "node_pool": {
          "type": "string",
          "title": "name of node pool"
        },
        "count": {
          "type": "integer",
          "format": "int32",
          "title": "total desired nodes"
        }
      }
    },
    "cmaaksScaleClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "wheather or not the cluster was scaled"
        },
        "status": {
          "type": "string",
          "title": "Status of the request"
        }
      }
    },
...
  },
...
}